6_Agents/2_Phidata_Reserch_Agent/requirements.txt
phidata 
openai 
duckduckgo-search 
newspaper4k 
lxml_html_clean
packaging
boto3